Dev Forcing sid to integer

+++ application/controllers/admin/Expressions.php
@@ -84,7 +84,7 @@ class Expressions extends SurveyCommonAction
     {
         $aData = array();
 
-        $sid = Yii::app()->request->getParam('sid', 0, 'integer');
+        $sid = (int) Yii::app()->request->getParam('sid', 0, 'integer');
         $surveyid = Yii::app()->request->getParam('surveyid', $sid, 'integer');
 
         $hasSurveyContentReadPermission = Permission::model()->hasSurveyPermission($sid, 'surveycontent', 'read');
